Rename HdfsConfigurationUpdater to HdfsConfigurationInitializer
